Actually, go back to variant definitions of imports/exports
Initially deemed an inconvenient form for validation, it's now proving
otherwise. For example, for validating exports we'll want to ensure that
all names are distinct, which is best handled by sorting a single list
and then doing comparisons of pairwise elements. Also, with the
broken-out forms, it made for repeated inefficient allocation
extensions during decoding. We'll break these out by type properly when
moving toward execution.